ok, some cleaning up and simplify on using integers for rotations in degree

+++ src/freert/library/LibPackagePin.java
@@ -31,12 +31,12 @@ public final class LibPackagePin implements java.io.Serializable
    // The location of the pin relative to its package. 
    public final PlaVectorInt relative_location;
    // the rotation of the pin padstack 
-   public final double rotation_in_degree;
+   public final int rotation_in_degree;
    
    /**
     * Creates a new package pin with the input coordinates relative to the package location.
     */
-   public LibPackagePin(String p_name, int p_padstack_no, PlaVectorInt p_relative_location, double p_rotation_in_degree)
+   public LibPackagePin(String p_name, int p_padstack_no, PlaVectorInt p_relative_location, int p_rotation_in_degree)
       {
       name = p_name;
       padstack_no = p_padstack_no;
